requirements.txt
pandas
streamlit
requests
streamlit-folium
folium
google-generativeai
